Add files via upload
Challenge 1:
A 3-tier environment is a common setup. Use a tool of your choosing/familiarity create these
resources on a cloud environment (Azure/AWS/GCP). Please remember we will not be judged
on the outcome but more focusing on the approach, style and reproducibility.

The solution has been made in AZURE Infrastructure , the architecture has been designed in powershell